requirements.txt
pyTelegramBotAPI>=4.14.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.2
matplotlib>=3.9.0
schedule>=1.2.1
psutil>=5.9.8
Flask>=3.0.0
Werkzeug>=3.0.1
Jinja2>=3.1.2
MarkupSafe>=2.1.3
boto3>=1.26.0